Add default static IP configuration and override mechanism
We don't want to use DHCP(?) for the time being, meaning that we
need a way to deploy the VM with a static IP address without having
to always change the source code.

Add the default IP (Dev environment) in config.ml, and allow these
values to be overridden through environment variables.